+++ tests/test_MTPviewer.py
@@ -26,7 +26,9 @@ from viewer.MTPclient import MTPclient
 from viewer.MTPviewer import MTPviewer
 from PyQt5.QtWidgets import QApplication
 from lib.rootdir import getrootdir
-from EOLpython.Qlogger.messageHandler import QLogger as logger
+from EOLpython.Qlogger.messageHandler import QLogger
+
+logger = QLogger("EOLlogger")
 
 
 class TESTMTPviewer(unittest.TestCase):
@@ -42,7 +44,7 @@ class TESTMTPviewer(unittest.TestCase):
         self.stream = StringIO()  # Set output stream to buffer
 
         # Instantiate a logger
-        self.log = logger.initLogger(self.stream, logging.INFO)
+        self.log = logger.initStream(self.stream, logging.INFO)
 
         # Location of default config file
         self.configfile = os.path.join(getrootdir(), 'Data', 'NGV',
@@ -65,7 +67,7 @@ class TESTMTPviewer(unittest.TestCase):
         self.viewer.clickBack()
         logger.flushHandler()
         self.assertRegex(self.stream.getvalue(),
-                         r"ERROR:.*No scans yet. Can't go backward\n")
+                         r".*ERROR | .*No scans yet. Can't go backward\n")
 
         # Test with JSON file
         filename = "../tests/test_data/DEEPWAVErf01.mtpRealTime.json"
@@ -76,7 +78,7 @@ class TESTMTPviewer(unittest.TestCase):
         self.viewer.clickBack()
         logger.flushHandler()
         self.assertRegex(self.stream.getvalue(),
-                         r"ERROR:.*On first scan. Can't go backward\n")
+                         r".*ERROR | .*On first scan. Can't go backward\n")
 
         # If not on last scan, should decrement viewScanIndex by one
         self.viewer.viewScanIndex = 3
@@ -90,7 +92,7 @@ class TESTMTPviewer(unittest.TestCase):
         self.viewer.clickFwd()
         logger.flushHandler()
         self.assertRegex(self.stream.getvalue(),
-                         r"ERROR:.*On latest scan. Can't go forward\n")
+                         r".*ERROR | .*On latest scan. Can't go forward\n")
 
         # Test with JSON file
         filename = "../tests/test_data/DEEPWAVErf01.mtpRealTime.json"
@@ -101,7 +103,7 @@ class TESTMTPviewer(unittest.TestCase):
         self.viewer.clickFwd()
         logger.flushHandler()
         self.assertRegex(self.stream.getvalue(),
-                         r"ERROR:.*On latest scan. Can't go forward\n")
+                         r".*ERROR | .*On latest scan. Can't go forward\n")
 
         # If not on last scan, should increment viewScanIndex by one
         self.viewer.viewScanIndex = 3