+++ tests/test_MTPviewer2.py
@@ -26,7 +26,9 @@ from viewer.MTPclient import MTPclient
 from viewer.MTPviewer import MTPviewer
 from PyQt5.QtWidgets import QApplication
 from lib.rootdir import getrootdir
-from EOLpython.Qlogger.messageHandler import QLogger as logger
+from EOLpython.Qlogger.messageHandler import QLogger
+
+logger = QLogger("EOLlogger")
 
 
 class TESTMTPviewer(unittest.TestCase):
@@ -42,7 +44,7 @@ class TESTMTPviewer(unittest.TestCase):
         self.stream = StringIO()  # Set output stream to buffer
 
         # Instantiate a logger
-        self.log = logger.initLogger(self.stream, logging.INFO)
+        self.log = logger.initStream(self.stream, logging.INFO)
 
         # Location of default config file
         self.configfile = os.path.join(getrootdir(), 'Data', 'NGV',
@@ -78,7 +80,7 @@ class TESTMTPviewer(unittest.TestCase):
         self.viewer.clickGo()
         logger.flushHandler()
         self.assertRegex(self.stream.getvalue(),
-                         r"ERROR:.*Select scan in range 1 - " +
+                         r".*ERROR | .*Select scan in range 1 - " +
                          str(self.viewer.currentScanIndex+1))
 
     def tearDown(self):