+++ tests/test_rcf.py
@@ -23,7 +23,9 @@ import unittest
 import logging
 from io import StringIO
 from util.rcf import RetrievalCoefficientFile
-from EOLpython.Qlogger.messageHandler import QLogger as logger
+from EOLpython.Qlogger.messageHandler import QLogger
+
+logger = QLogger("EOLlogger")
 
 
 class TESTrcf(unittest.TestCase):
@@ -45,7 +47,7 @@ class TESTrcf(unittest.TestCase):
         self.stream = StringIO()  # Set output stream to buffer
 
         # Instantiate a logger
-        self.log = logger.initLogger(self.stream, logging.INFO)
+        self.log = logger.initStream(self.stream, logging.INFO)
 
     def testRCFId(self):
         """ Test that the RCF Id is accurately parsed from the filename """