+++ tests/test_rcf_set.py
@@ -26,7 +26,9 @@ from util.rcf_set import RetrievalCoefficientFileSet
 
 import logging
 from io import StringIO
-from EOLpython.logger.messageHandler import Logger as logger
+from EOLpython.logger.messageHandler import Logger
+
+logger = Logger("EOLlogger")
 
 
 class TESTrcfSet(unittest.TestCase):
@@ -40,7 +42,7 @@ class TESTrcfSet(unittest.TestCase):
         # Set up logging
         self.stream = StringIO()  # Set output stream to buffer
         loglevel = logging.INFO
-        logger.initLogger(self.stream, loglevel)
+        logger.initStream(self.stream, loglevel)
 
         self.Directory = "../tests/test_data"
         # Constants the apply specifically to the CSET RCF file in the dir
@@ -134,7 +136,7 @@ class TESTrcfSet(unittest.TestCase):
             # Test that user was shown appropriate error message
             logger.flushHandler()
             self.assertRegex(self.stream.getvalue(),
-                             "ERROR:.*Failed to make fileset. Requested " +
+                             ".*ERROR | .*Failed to make fileset. Requested " +
                              "RCF file NRCDA067 does not exist in RCFdir" +
                              " ../tests/test_data/RC")