+++ tests/test_readascii_parms.py
@@ -22,7 +22,9 @@ from util.readascii_parms import AsciiParms
 
 import sys
 import logging
-from EOLpython.logger.messageHandler import Logger as logger
+from EOLpython.logger.messageHandler import Logger
+
+logger = Logger("EOLlogger")
 
 
 class TESTreadascii_parms(unittest.TestCase):
@@ -31,7 +33,7 @@ class TESTreadascii_parms(unittest.TestCase):
         # Setup logging
         self.stream = sys.stdout  # Send log messages to stdout
         loglevel = logging.INFO
-        logger.initLogger(self.stream, loglevel)
+        logger.initStream(self.stream, loglevel)
 
         # Set environment var to indicate we are in testing mode
         os.environ["TEST_FLAG"] = "true"